+++ packages/hyperverse-evm-erc777/source/erc777Library.ts
@@ -73,7 +73,7 @@ export async function ERC777LibraryInternal(
 	const getTotalSuply = async () => {
 		try {
 			const totalSupply = await base.proxyContract?.totalSupply();
-			return BigNumber.from(totalSupply) as BigNumber;
+			return BigNumber.from(totalSupply).toNumber();
 		} catch (error) {
 			throw error;
 		}
@@ -82,16 +82,7 @@ export async function ERC777LibraryInternal(
 	const getBalanceOf = async (account: string) => {
 		try {
 			const balance = await base.proxyContract?.balanceOf(account);
-			return BigNumber.from(balance) as BigNumber;
-		} catch (error) {
-			throw error;
-		}
-	};
-
-	const getBalance = async () => {
-		try {
-			const balance = await base.proxyContract?.balance();
-			return BigNumber.from(balance) as BigNumber;
+			return BigNumber.from(balance).toNumber();
 		} catch (error) {
 			throw error;
 		}
@@ -139,7 +130,7 @@ export async function ERC777LibraryInternal(
 	const allowance = async (owner: string, spender: string) => {
 		try {
 			const allowance = await base.proxyContract?.allowance(owner, spender);
-			return BigNumber.from(allowance) as BigNumber;
+			return BigNumber.from(allowance).toNumber();
 		} catch (error) {
 			throw error;
 		}
@@ -275,7 +266,6 @@ export async function ERC777LibraryInternal(
 		getDecimal,
 		getGranularity,
 		getTotalSuply,
-		getBalance,
 		getBalanceOf,
 		send,
 		transfer,