+++ packages/hyperverse-evm-erc777/stories/allowance.stories.tsx
@@ -22,6 +22,6 @@ const Template = (args) => (
 export const Demo = Template.bind({});
 
 Demo.args = {
-	owner: '',
-	spender: ''
+	owner: '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266',
+	spender: '0x976EA74026E726554dB657fA54763abd0C3a0aa9',
 };