+++ packages/hyperverse-evm-erc777/stories/send.stories.tsx
@@ -21,8 +21,10 @@ const Template = (args) => (
 
 export const Demo = Template.bind({});
 
+const { ethers } = require('hardhat');
+
 Demo.args = {
-	to: '',
+	to: '0x976EA74026E726554dB657fA54763abd0C3a0aa9',
 	amount: 10000,
-	data: '',
+	data: ethers.utils.formatBytes32String('0x'),
 };