+++ packages/hyperverse-evm-tribes/stories/utils/Provider.tsx
@@ -1,7 +1,7 @@
 import { initialize, Network, NetworkConfig, Provider } from '@decentology/hyperverse';
 import { Localhost, Ethereum } from '@decentology/hyperverse-evm';
 import { FC, VFC } from 'react';
-import * as Tribes from '../../source';
+import * as ERC721 from '../../source';
 
 export const HyperverseProvider: FC<{}> = ({ children }) => {
 	const hyperverse = initialize({
@@ -18,11 +18,19 @@ export const HyperverseProvider: FC<{}> = ({ children }) => {
 				  }
 				: {
 						type: Network.Testnet,
-						chainId: 1337,
+						chainId: 31337,
 						name: 'localhost',
 						networkUrl: 'http://localhost:6006/hyperchain',
 				  },
-		modules: [{ bundle: Tribes, tenantId: '0x62a7aa79a52591Ccc62B71729329A80a666fA50f' }],
+		modules: [
+			{
+				bundle: ERC721,
+				tenantId:
+					process.env.STORYBOOK_NETWORK === 'rinkeby'
+						? '0x62a7aa79a52591Ccc62B71729329A80a666fA50f'
+						: '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266',
+			},
+		],
 	});
 	return <Provider initialState={hyperverse}>{children}</Provider>;
 };