+++ apps/ethereum/stake-rewards/CHANGELOG.md
@@ -1,5 +1,17 @@
 # token
 
+## 0.1.2
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @decentology/config@1.0.3
+    -   @decentology/hyperverse@1.0.15
+    -   @decentology/hyperverse-ethereum@1.0.15
+    -   @decentology/hyperverse-evm-erc777@2.0.4
+    -   @decentology/hyperverse-evm-stake-rewards@1.0.2
+    -   @decentology/unstated-next@1.1.9
+
 ## 0.1.1
 
 ### Patch Changes