+++ packages/hyperverse-ethereum-randompick/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @decentology/hyperverse-ethereum-randompick
 
+## 1.0.17
+
+### Patch Changes
+
+-   Milestone version
+-   Updated dependencies
+    -   @decentology/hyperverse@1.0.15
+    -   @decentology/hyperverse-evm@1.0.15
+    -   @decentology/hyperverse-storage-skynet@1.0.14
+    -   @decentology/unstated-next@1.1.9
+
 ## 1.0.16
 
 ### Patch Changes